requirements.txt
fastapi
uvicorn
pytesseract
pillow
streamlit
transformers
torch
requests

pix2tex[api]